requirements.txt
beautifulsoup4==4.9.1
censys==0.0.8
cmd2==0.8.9
csirtgsdk==1.1.5
cymon==0.1
dnspython==1.15.0
dshield==0.2.1
feedparser==5.2.1
fuzzyfinder==2.1.0
ipwhois==1.0.0
OTXv2==1.5.10
pygments==2.6.1
pymongo==3.5.1
python-libnmap==0.7.0
python-whois==0.6.9
redis==2.10.6
requests==2.20.0
threatcrowd==0.4
yara==1.7.7
